+++ src/assets/catalog/base_data/colors.js
@@ -1,4 +1,4 @@
-export let color = {
+export const color = {
   0: {
     0: { h: 0, s: 0, l: 100, hex: "#ffffff", name: "White" },
     1: { h: 0, s: 0, l: 83, hex: "#d4d4d4", name: "Smoke" },
