+++ src/assets/catalog/base_data/genders.js
@@ -1,0 +1,33 @@
+export const genders = {
+    phisiological: {
+        male: "male",
+        intersex: "intersex",
+        female: "female",
+        demiboy: "demiboy",
+        demigirl: "demigirl",
+        neutral: "neutral_gender",
+        unstable: "unstable",
+        other: "other_gender"
+    },
+
+    feel: {
+        androgin:"androgin",
+        pangender:"pangender",
+        transsexual:"transsexual",
+        demigender:"demigender",
+        gender_fluid:"gender_fluid",
+        cisgender:"cisgender",
+        undecided:"undecided",
+        queer:"queer",
+        neutral: "neutral_gender",
+        bigender: "bigender",
+        other:"other_gender"
+    },
+
+    look: {
+        as_phisiological: "as_phisiological",
+        as_feel: "as_feel",
+        as_other:"as_other_gender"
+    }
+
+}
